pyproject.toml
[project]
name = "iac-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.23.0",
    "gitpython>=3.1.44",
    "langchain-openai>=0.3.23",
    "mcp[cli]>=1.9.4",
    "mcp-cli[cli,dev,server]>=0.1.0",
    "openai>=1.88.0",
    "python-hcl2>=7.2.1",
    "uvicorn>=0.34.3",
]
